+++ src/constants.js
@@ -1,4 +1,4 @@
 export const DefaultAppName = 'app';
 export const TemplateExt = '.mustache';
 export const ManifestoFileName = 'manifesto.json';
-export const Directive = { Copy: 'copy', Loop: 'loop' };
+export const Directive = { Copy: 'copy', Loop: 'loop', ViewModel: 'viewModel' };